Add memory context switching for sequential r/w
Several tasks might use memory simultainously. This is problematic
for sequential read/write sessions where a read or write operation
from another task changes the address latch values.

This moves the latch mirrors out of the memory module and into the
caller, and adds the necessary logic to check whether the context
has changed, in which case the latch values will be updated with
the new context, so that sequential operations can be continue.