New battery backup circuit
The DS1210 is expensive and found to not really be needed here, so it has been
replaced by a discrete transistor circuit.

The op-amp used for the bias voltage buffer in the audio amplifier is now used
for buffering the battery voltage instead. Its output is now connected to ADC5
of the ATmega to allow for reading the battery voltage.