Main board v0.9: fix memory backup, new MIDI connector, BOM fixes
The discrete memory backup circuit turned out to be too fragile. The old
DS1210 based design is back, but it is now also used to provide the CPU
RESET signal to further reduce the possibility of memory corruption. The
battery voltage is still being connected (through a buffer) to the
Atmega for being able to check the battery voltage.

The MIDI connector that was used previously is not being stocked at
Mouser any more, so it has been replaced by one that is.

A couple of fixes have been done to the BOM, replacing out of stock
components with drop-in replacements (a resistor and the 74'238).

This should hopefully be the last revision of the main board...