mute channels at startup
Fix the annoying issue where square and noise channels tend to
be noisy at startup until something triggers a note to be played.
Not sure if this is an issue for all 2A03 versions/clones.

The way this is done is by spinning up a new task which does this once
at startup from quite a high level. Could probably be fixed with lower
level register writes, but this utilizes the existing architecture.