main: improve battery check
Wait a bit before reading initial battery measurement.
If this is not done the read value is close to 0 and will cause a
battery level warning even if the battery is OK.

Also remove offset adjustment in the battery_read function - this is not
necessary with the current circuit design.