fix bug in memory initialization
The DDRC register was written before the PORTC register when setting WE
high in the setup_memory function. This causes WE to be driven low for a
short period of time. At the point in time when this happens, the CE1
and CE2 chip enable signals for the RAMs are still being driven low, so
this actually causes a write to the memory.

Make sure that the chip enable signals are driven high before setting
WE, and also set WE with the correct sequence where PORTC is written
before DDRC.

With this fixed, put the magic words back to start at address 0.