sample: make sure cached block state does not overwrite end-of-chain flag
Fix a sneaky bug where the block state tree update part of
allocate_block overwrites the newly set end-of-chain flag if the new
block index is the same as one of the tree node indices.
This would cause corrupt chains and cause trouble when deleting samples
and then loading new ones.