Support booting without 2A03 installed
Support booting the NESizer without a 2A03 chip installed.

To check for 2A03 presence, the interrupt on the PC3 (R/W) pin is
enabled and used to count toggles on the line. Testing shows that on
average a regular 2A03 (divisor 12) will cause the count to become ~110
during a 10 us window. This value will be lower for chips with higher
clock divisors, so the threshold is set to a count of 10 to keep a good
margin. Since the pin will have an internal pull-up when configured as
input, the test could probably have been that the count is non-zero, but
to avoid any potential glitching when enabling the interrupt, checking
for a value a bit higher than that made sense.

The settings UI when pressing 'CHIP TYPE' will show '--' in the 7
segment display if no 2A03 was detected.  The functions for talking to
the 2A03 are replaced with empty variants in this case (also if 2A03
presence is detected, but the chip type detection failed).